+++ src/main/java/co/modularbank/banking/controller/model/SaveTransactionResponse.java
@@ -1,20 +1,22 @@
 package co.modularbank.banking.controller.model;
 
+import java.math.BigDecimal;
+
 public class SaveTransactionResponse extends TransactionResponse {
-    private double balance;
+    private BigDecimal balance;
 
     public SaveTransactionResponse() {}
 
-    public SaveTransactionResponse(long accountId, long transactionId, double amount, String currency, String direction, String description, double balance) {
+    public SaveTransactionResponse(long accountId, long transactionId, BigDecimal amount, String currency, String direction, String description, BigDecimal balance) {
         super(accountId, transactionId, amount, currency, direction, description);
         this.balance = balance;
     }
 
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 
-    public void setBalance(double balance) {
+    public void setBalance(BigDecimal balance) {
         this.balance = balance;
     }
